requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
PyPDF2
python-docx
google-generativeai
python-jose
passlib
bcrypt
python-multipart
httpx